refactor current database code into its own service
